requirements.txt
asgiref==3.4.1
Django==3.2.9
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.9.0
importlib-metadata==4.8.2
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
zipp==3.6.0
